requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
dataclasses-json>=0.5.7
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0